+++ examples/resources/timescale_log_exporter/resource.tf
@@ -2,7 +2,7 @@ terraform {
   required_providers {
     timescale = {
       source  = "timescale/timescale"
-      version = "~> 2.10"
+      version = "~> 2.13"
     }
   }
 }